getUserTokens for native and erc20
